Combined separate modules into a single module
This combines the methods developed into a single module for use
in AWS Lambda.

Also updated unit tests to match new method locations.
Added support for overriding configuration defaults with environment
variables.
Added a policy constraint when assuming roles for safety.
Added some logging to make it easier to follow in Lambda
